Added Solution for Coin Change Problem in C++
The idea is to use recursion to solve the Coin Change Problem problem. We recur to see if the total can be reached by choosing the coin or not for each coin of given denominations. If choosing the current coin results in the solution, update the total number of ways.